refactor(embed): move buildChatTheme to config.ts
Decouple inline-chat and floating-chat. buildChatTheme is a pure
EmbedConfig → ChatTheme adapter with no rendering concerns, so it
belongs alongside resolveConfig in config.ts rather than being
imported across peer component files.
